Day 01
Iterate over each line of the input. Split the line by a three space
delimiter/separator. Take the first value and append it to a left array.
Take the second and append it to the right array. After iterating over
every line, sort the left and right arrays. Return the result as a
tuple.

Part 1
Take the left and right value at an index, take the
absolute value of the difference, then add that value to a running sum.

Part 2
Iterate over the array of right values and create a fequency map. The
key is the value and the value is the number of times that value appears
in the right array. Then we iterate over the left array and multiply the
value by the frequency it appeared in the right array. We add that
result to the running sum.