Days 02 & 03
Day 02

Iterate over each line of the input. Split the line by a single space
delimiter/separtor. Map this into an array of Integers. The output of
data ingest will be an array of integer arrays.

Part 1
Each row of entities is a report. Each report has levels (the Integers).
We want to count the number of safe reports. For a report to be "safe",
the following must be true:
- Levels are either all increasing or all decreasing.
- Any two adjacent levels will differ by at least one and at most three.

Part 2
We want to count the number of safe reports again, but this time we
account for the dampener. The dampener means that a single level could
be removed from a report and if the report is then safe, it's considered
safe.

Day 03

Read multiplication commands from the input. This is accomplished with a
Regular Expression.

Part 1
Sum the product of all the multiplication commands.

Part 2
Sum the product of all the multiplication commands, but this time we
need to account for the `do()` and `don't()` commands that essentially
enable/disable the following multiplication commands.