Days 04 & 05
Day 04

Each line of input will become an array of Characters.

Part 1
Word search. We're looking for "XMAS". The phrase can be horizontal,
vertical, or diagonal. The phrase can be forwards ("XMAS") or backwards
("SAMX").

My naive approach is to find an X, then look in the compass directions:
North, Northeast, East, Southeast, South, Southwest, West, or Northwest.
This is not very efficient and takes a while to run. It seems like it
took somewhere in the 30-60s ballpark on my M3 iMac. I also think that
this implementation used a lot of copy/paste for its implementation and
is a bit sloppy.

Part 2
Still a word search, but instead of looking for "XMAS", we are looking
for an "X" pattern where MAS is diagonal.

My approach here is to look for an A, then look for MAS or SAM in the
diagonal. This ran much faster than Part 1.

Day 05

Split on a "\n\n" delimiter. The first half is the rules, which we
separate by "|" into a left and right part. We parse this into a
dictionary where the left number is the key and the right number goes
into a Set. The second half is the "jobs", which is an array of
integers. I stored this into a proper model.

Part 1
Find the "valid" jobs, where the job is sorted by the given rules. We
sum the median from the valid jobs to get the answer.

Part 2
Find the "invalid" jobs (the jobs that weren't valid) and we sort them
so that they become valid (dictated by the rule set). We then sum the
median from the newly sorted jobs to get the answer.

This solution ran slower than Part 1. I'm assuming it's the sorting
portion that took some amount of time. Maybe 10-15s.